8326653: Remove jdk.internal.reflect.UTF8
